pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "A powerful universal web scraping and content analysis system based on MCP (Model Context Protocol)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepseek>=1.0.0",
    "httpx>=0.28.1",
    "mcp>=1.6.0",
    "openai>=1.70.0",
    "openrouter>=1.0",
    "python-dotenv>=1.1.0",
    "ruff>=0.9.4",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = ["web/**", "writers/**", "server.py"]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E402", "E501"]

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
ignore_missing_imports = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false
exclude = "(server\\.py|.*html_cleaner\\.py|.*playwright_fetcher\\.py|.*crawler_framework\\.py)"
explicit_package_bases = true
mypy_path = "."
no_implicit_optional = false
check_untyped_defs = false
[[tool.mypy.overrides]]
module = "utils.html_cleaner"
ignore_errors = true

[[tool.mypy.overrides]]
module = "utils.playwright_fetcher"
ignore_errors = true

[[tool.mypy.overrides]]
module = "utils.crawler_framework"
ignore_errors = true

[[tool.mypy.overrides]]
module = "scripts.universal_knowledge_processor"
ignore_errors = true
